pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry]
authors = ["Martin Hjelmare <marhje52@gmail.com>"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Topic :: Home Automation",
]
description = "Python asyncio package to connect to MySensors gateways."
license = "Apache Software License 2.0"
name = "aiomysensors"
packages = [
  {include = "aiomysensors", from = "src"},
]
readme = "README.md"
repository = "https://github.com/MartinHjelmare/aiomysensors"
version = "0.3.19"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/MartinHjelmare/aiomysensors/issues"
"Changelog" = "https://github.com/MartinHjelmare/aiomysensors/blob/main/CHANGELOG.md"

[tool.poetry.scripts]
aiomysensors = 'aiomysensors.cli:cli'

[tool.poetry.dependencies]
aiofiles = "^24.0.0"
asyncio-mqtt = "^0.16"
awesomeversion = "^24.0.0"
marshmallow = "^3.17"
pyserial-asyncio = "^0.6"
python = "^3.12"
rich = ">=10"
typer = {extras = ["all"], version = "^0.9.0"}

[tool.poetry.group.dev.dependencies]
copier = "^9.3.1"
mypy = "^1.0"
pre-commit = "^3.0.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
python-semantic-release = "^9.0.0"
ruff = "^0.6.6"
types-aiofiles = "^22.0.0"

[tool.ruff]
line-length = 88
lint.ignore = [
  "COM812", # missing-trailing-comma
  "D203", # 1 blank line required before class docstring
  "D213", # Multi-line docstring summary should start at the second line
  "D417", # False positives in some occasions
  "EM101", # raw-string-in-exception
  "EM102", # f-string-in-exception
  "ISC001", # single-line-implicit-string-concatenation
  "PD011", # pandas-use-of-dot-values
  "PLR0913", # too-many-arguments
  "PLR2004", # magic-value-comparison
  "TCH001", # typing-only-first-party-import
  "TCH002", # typing-only-third-party-import
  "TCH003", # typing-only-stdlib-import
  "TRY003", # raise-vanilla-args
]
lint.isort.force-sort-within-sections = true
lint.isort.known-first-party = ["aiomysensors", "tests"]
lint.per-file-ignores."setup.py" = ["D100"]
lint.per-file-ignores."tests/**/*" = [
  "FBT001", # boolean-type-hint-positional-argument
  "S101", # assert
]
lint.select = [
  "ALL", # all rules
]
target-version = "py312"

[tool.pytest.ini_options]
addopts = """\
    -Wdefault
    --cov=aiomysensors
    --cov-report=term-missing
    --cov-report=xml
    """
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "@overload",
  "def __repr__",
  "if TYPE_CHECKING",
  "raise AssertionError",
  "raise NotImplementedError",
  "raise RuntimeError",
  'if __name__ == "__main__":',
]

[tool.mypy]
check_untyped_defs = true
#disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
enable_incomplete_feature = "NewGenericSyntax"
exclude = [
  'setup.py',
]
mypy_path = "src/"
no_implicit_optional = true
show_error_codes = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
allow_untyped_defs = true
module = "tests.*"

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
  "paho",
  "paho.mqtt",
  "paho.mqtt.client",
  "serial_asyncio",
]

[tool.semantic_release]
build_command = "pip install poetry && poetry build"
major_on_zero = false
version_toml = ["pyproject.toml:tool.poetry.version"]
version_variables = [
  "src/aiomysensors/__init__.py:__version__",
]

[tool.semantic_release.changelog]
exclude_commit_patterns = [
  "chore.*",
  "ci.*",
  "Merge pull request .*",
]

[tool.semantic_release.changelog.environment]
keep_trailing_newline = true

[tool.semantic_release.branches.main]
match = "main"

[tool.semantic_release.branches.noop]
match = "(?!main$)"
prerelease = true
